+++ src/main/kotlin/com/dansplugins/factionsystem/service/Services.kt
@@ -3,12 +3,12 @@ package com.dansplugins.factionsystem.service
 import com.dansplugins.factionsystem.chat.MfChatService
 import com.dansplugins.factionsystem.claim.MfClaimService
 import com.dansplugins.factionsystem.duel.MfDuelService
-import com.dansplugins.factionsystem.dynmap.MapService
 import com.dansplugins.factionsystem.faction.MfFactionService
 import com.dansplugins.factionsystem.gate.MfGateService
 import com.dansplugins.factionsystem.interaction.MfInteractionService
 import com.dansplugins.factionsystem.law.MfLawService
 import com.dansplugins.factionsystem.locks.MfLockService
+import com.dansplugins.factionsystem.map.MapService
 import com.dansplugins.factionsystem.notification.MfNotificationService
 import com.dansplugins.factionsystem.player.MfPlayerService
 import com.dansplugins.factionsystem.potion.MfPotionService
